* hipl-dev@freelists.org--hipl/hipl--userspace--2.6--patch-1627
   A bug fix to lsi/userspace ipsec output

 * hipl-dev@freelists.org--hipl/hipl--userspace--2.6--patch-1628
   A bug fix to lsi/userspace ipsec output

 * hipl-dev@freelists.org--hipl/hipl--userspace--2.6--patch-1629
   A bug fix to lsi/userspace ipsec output

 * hipl-dev@freelists.org--hipl/hipl--userspace--2.6--patch-1630
   Some commenting in firewall

 * hipl-dev@freelists.org--hipl/hipl--userspace--2.6--patch-1631
   Commenting

 * hipl-dev@freelists.org--hipl/hipl--userspace--2.6--patch-1632
   Changed some 'int' variables to uint16_t from userspace ipsec

 * hipl-dev@freelists.org--hipl/hipl--userspace--2.6--patch-1633
   Bug fixes on openwrt code merged in patch-1624

+++ firewall/esp_prot.h
@@ -7,7 +7,7 @@
 int esp_prot_init(void);
 int esp_prot_set_sadb(hip_sa_entry_t *entry, uint8_t esp_prot_transform,
 		unsigned char *esp_prot_anchor, int direction);
-int add_esp_prot_hash(unsigned char *out_hash, int *out_length, hip_sa_entry_t *entry);
+int add_esp_prot_hash(unsigned char *out_hash, uint16_t *out_length, hip_sa_entry_t *entry);
 int verify_esp_prot_hash(hip_sa_entry_t *entry, unsigned char *hash_value);
 hash_chain_t * esp_prot_get_corresponding_hchain(unsigned char *hchain_anchor,
 		uint8_t transform);
